Implemented various fixes and improvements: - Fixed issues with success, fail, and error handling throughout the project. - Ensured proper functionality of the cart and profile pages. - Improved responsiveness across the entire project, including the home page, header, and footer components. - Updated and added necessary data for improved user experience. - Incorporated a developers' contact page in the footer section. - Activated links in the footer for seamless navigation. - Resolved backend issue preventing re-purchasing of products. - Added controller functionality to view purchased orders upon login.

+++ Backend/src/main/java/com/ecom/controller/ProductController.java
@@ -82,7 +82,7 @@ public class ProductController {
 	// not tested 
 	@DeleteMapping("/deleteproduct/{productId}")
 	public boolean deleteProductById(@PathVariable int productId) {
-		return productService.deleteProductById(productId);
+		return productService.deleteProductById((long) productId);
 	}
 
 	// not tested
